requirements.txt
Flask==3.0.3
pymongo==4.8.0
dnspython==2.6.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-search-results==2.4.2=